reload: use the global startup config instance.
Prior to this commit `reload.py` unconditionally parsed the hardcoded
`~/.ldmud-efuns` config file to figure out which efuns are enabled for
reload purposes.

Since `startup.py` now both offers a way for consumers to specify
a _different_ config file, and exposes the config that it loaded as
a global we need to rework the reload code.

With this commit the reload code uses the global config read by
`startup.py` instead of reading its own.