startup/reload: reload INI config before efuns.
To be able to enable/disable efuns after the configuration has been
loaded by invoking the `python_reload` efun we need to ensure the reload
logic _also_ reloads the INI configuration.

This is slightly complicated by the fact that we now allow `startup.py`
to be used with custom INI filepaths.

The solution in this commit is to have `startup.py` remember the file
name used to parse its config, and offer a function to reload that same
file that `reload.py` can use.